logger: refactor integration
* Does not monkey-patch the Logger class anymore and uses the Decorator
  pattern now
* Airbrake severity level cannot be set to DEBUG or INFO to avoid
  massive spam. When that happens our own airbrake-ruby spams the
  dashboard with requests continuously, which leads to DDoS'ing our
  service. We can technically stop logging every response in
  airbrake-ruby, but this is still very dangerous because any other
  library that we do not control can do the same without even realising
  it.
