+++ README.md
@@ -417,44 +417,44 @@ end
 
 If you want to convert your log messages to Airbrake errors, you can use our
 integration with Ruby's `Logger` class from stdlib. All you need to do is to
-make sure that the integration is required (if the Airbrake is loaded after
-`Logger`, then it is already integrated):
+wrap your logger in Airbrake's decorator class:
 
 ```ruby
-require 'airbrake/logger/logger_ext'
+# Create a normal logger
+logger = Logger.new(STDOUT)
+
+# Wrap it
+logger = Airbrake::AirbrakeLogger.new(logger)
 ```
 
-To start using the integration you don't have to perform any additional steps,
-simply use Logger as you normally do. Messages will be both logged and sent to
-Airbrake:
+Now you can use the `logger` object exactly the same way you use it. For
+example, calling `fatal` on it will both log your message and send it to the
+Airbrake dashboard:
 
-```ruby
-logger = Logger.new(STDOUT)
+```
 logger.fatal('oops')
 ```
 
-The Logger class will attempt to utilize the default Airbrake notifier to send
-messages. It's possible to set the notifier explicitly:
+The Logger class will attempt to utilize the default Airbrake notifier to
+deliver messages. It's possible to redefine it via `#airbrake_notifier`:
 
 ```ruby
-# New method.
-# We assign the default notifier here.
-logger.airbrake_notifier = Airbrake[:default]
+# Assign your own notifier.
+logger.airbrake_notifier = Airbrake[:other_notifier]
 ```
 
 #### Airbrake severity level
 
-In order to reduce the noise from the Logger integration (or maximize) it's
-possible to configure Airbrake severity level. For example, if only want to send
-fatal messages from Logger, then configure your logger as follows:
+In order to reduce the noise from the Logger integration it's possible to
+configure Airbrake severity level. For example, if you want to send only fatal
+messages from Logger, then configure it as follows:
 
 ```ruby
-# New method.
-# Send only fatal messages to Airbrake.
-logger.airbrake_severity_level = Logger::FATAL
+# Send only fatal messages to Airbrake, ignore anything below this level.
+logger.airbrake_level = Logger::FATAL
 ```
 
-By default, `airbrake_severity_level` is set to `Logger::WARN`, which means it
+By default, `airbrake_level` is set to `Logger::WARN`, which means it
 sends warnings, errors and fatal error messages to Airbrake.
 
 ### Plain Ruby scripts
